Refactor and add default options
A minor refactor changing the way URLs are formed internally and adding default
options for upload and download. In addition to bringing the SDK more in line
with the other ones, this refactor makes it easier to implement more endpoint
functions s well as to add encryption to uploads and downloads.