[BREAKING] Fix test NDFs by using server URL instead of portal URL
This change helps prevent cases where you would try to access data right after
uploading it and get routed to a different server by the load balancer. Now, you
will always be taken to the same server for the duration of the client's
lifetime.

**NOTE:** This is a breaking change because it changes the URLs generated by the
SDK and changes the return of `client.portalURL()`. For example, we probably
want skybin links to remain as `siasky.net/<skylink>` instead of
`us-va-1.siasky.net/<skylink>`. We may have to think this through more.