[path-seed] Only allow directory seeds for `deriveEncryptedPathSeed`
This commit reverts `deriveEncryptedPathSeed` to only accepting 64-byte
directory path seed inputs.

In a previous commit we enabled 32-byte file path seed inputs to
`deriveEncryptedPathSeed` so that MySky could pass the root path seed into this
SDK function.

(The root path seed is 32 bytes despite being a directory path seed, and must
remain that way. It cannot change or all path seeds will break since all seeds
are derived from the root seed.)

The problem was that 32-byte file path seeds should not be accepted by
`deriveEncryptedPathSeed`, only 64-byte directory seeds.

Various solutions were proposed and discussed, but we decided to simply
copy/paste this function to MySky and change the the validation to only accept a
32-byte root seed. The implementation for `deriveEncryptedPathSeed` is unlikely
to change as doing so would break compatibility, so copy/pasting for internal
use in MySky is fairly safe.

This commit therefore reverts `deriveEncryptedPathSeed` to only accepting
64-byte directory path seed inputs.