[Breaking] Use `sign.detached` instead of `sign` to get signature
Previously, the returned signature was longer than `SIGNATURE_LENGTH`. This is
because `sign` returns the signed message (including the original message) which
was not necessary -- we only need it to return the signature.

In addition to being unnecessary, it was also confusing to have a signature that
was longer than expected.

We should have been using `sign.detached`. We already use `verify.detached`.