Don't throw if portal redirect fails
We don't want to fail to initialize MySky or to login if the portal redirect
fails. This can happen if the user entered an invalid portal, and in that case
he wouldn't be able to log in to change the portal.

Also, **this PR temporarily disables the portal redirect** until the autologin
changes in MySky are deployed.

Once this is merged in the `master` branch, we can merge `master` into the
`beta` branch and release a new beta version which includes SkyDB V2 and has
parallel uploads re-enabled. (We can also start working on a new stable
release.)

This commit was confirmed to be working by testing at:
https://bjftds.hns.siasky.net/

The relevant PR for the test skapp can be found at:
https://github.com/SkynetLabs/test-skapp/pull/7