Make `ExecuteRequestError` backwards-compatible with `AxiosError`
We previously made a change where we were returning a new error type,
`ExecuteRequestError`, from failed network requests instead of `AxiosError`. The
original Axios error was retrievable with `error.originalError`. This was a
break in compatibility for third-party devs who may have come to rely on the
returned `AxiosError`.

This commit updates `ExecuteRequestError` to be fully compatible with
`AxiosError`. Now this won't be a breaking change.